move properties related files back to endpoints package
