+++ core/src/main/java/com/linecorp/armeria/client/endpoint/RestartableThread.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.client.endpoint.properties;
+package com.linecorp.armeria.client.endpoint;
 
 import static com.google.common.base.Preconditions.checkState;
